Move Cloud Run worker-ID plugin under cloud_run/worker_id/
Relocate the Google Cloud Run worker-ID plugin from directly inside
temporalio/contrib/gcp/cloud_run/ into a new worker_id/ sub-package so it
owns its own __init__.py and README.md. This avoids a hard collision with
the separate OTel Cloud Run plugin, which also owns cloud_run/README.md and
cloud_run/__init__.py; after the move the two plugins share only the
minimal cloud_run/ and gcp/ namespace-marker __init__.py files.

The public names are unchanged; only the import path gains .worker_id:

    from temporalio.contrib.gcp.cloud_run.worker_id import WorkerIDPlugin

cloud_run/__init__.py is reduced to a minimal namespace-marker docstring
with no worker-ID exports. Mirrors the Go (contrib/gcp/cloudrun/workerid)
and .NET (CloudRun.WorkerId) layouts.
