+++ temporalio/contrib/gcp/cloud_run/worker_id/_metadata.py
